requirements.txt
# Core Dependencies
Django==4.2.7
psycopg2==2.9.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
# psycopg2-binary==2.9.9
# gunicorn==21.2.0

# # Authentication
# django-allauth==0.58.2
# djangorestframework-simplejwt==5.3.0
# django-cors-headers==4.3.0

# # Database and Caching
# redis==5.0.1
# django-redis==5.4.0
# celery==5.3.4

# # API Documentation
# drf-yasg==1.21.7
# django-filter==23.3.0
# Markdown==3.5.1

# # Image Processing
# Pillow==10.1.0

# # Development & Debug Tools
# django-debug-toolbar==4.2.0
# ipython==8.12.0
# Werkzeug==3.0.1

# # Testing
# pytest==7.4.3
# pytest-django==4.7.0
# coverage==7.3.2
# factory-boy==3.3.0

# # Code Quality
# black==23.11.0
# flake8==6.1.0
# isort==5.12.0
# mypy==1.7.0

# # Security
# django-environ==0.11.2
# django-storages==1.14.2

# # AWS
# boto3==1.29.3
# botocore==1.32.3

# # Production
# sentry-sdk==1.32.0
# django-anymail==10.1

# # Export/Import
# openpyxl==3.1.2
# pandas==2.1.3
# xlrd==2.0.1

# # Background Tasks
# django-q==1.3.9
# django-background-tasks==1.2.5

# # Optimization
# django-queryset-csv==1.1.0
# django-extensions==3.2.3
# django-optimized-image==0.3.0

# # Internationalization
# django-modeltranslation==0.18.11
# Babel==2.13.1